requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
openai==1.3.0
pytz==2023.3
